refactor: detach dependencies to chatting from socket
