+++ src/chatting/chatting.module.ts
@@ -1,15 +1,16 @@
 import { AgoraModule } from '../agora/agora.module';
 import { dynamooseModule } from '../config.dynamoose';
-import { SocketModule } from '../socket/socket.module';
+import { FcmModule } from '../fcm/fcm.module';
+import { RedisModule } from '../redis/redis.module';
 import { UserModule } from '../user/user.module';
 import { ChattingRepository } from './chatting.repository';
 import { ChattingService } from './chatting.service';
 import { Module } from '@nestjs/common';
 
 @Module({
-  imports: [dynamooseModule, AgoraModule, UserModule, SocketModule],
+  imports: [dynamooseModule, AgoraModule, UserModule, RedisModule, FcmModule],
   controllers: [],
   providers: [ChattingService, ChattingRepository],
-  exports: [ChattingRepository],
+  exports: [ChattingRepository, ChattingService],
 })
 export class ChattingModule {}
