+++ src/offer/offer.service.ts
@@ -1,4 +1,5 @@
 import { ChattingRepository } from '../chatting/chatting.repository';
+import { ChattingService } from '../chatting/chatting.service';
 import { ChattingStatus } from '../chatting/entities/chatting.interface';
 import { QuestionRepository } from '../question/question.repository';
 import { Fail, Success } from '../response';
@@ -15,6 +16,7 @@ export class OfferService {
     private readonly questionRepository: QuestionRepository,
     private readonly tutoringRepository: TutoringRepository,
     private readonly socketService: SocketService,
+    private readonly chattingService: ChattingService,
   ) {}
 
   async append(userId: string, questionId: string) {
@@ -54,14 +56,14 @@ export class OfferService {
         text: '안녕하세요 선생님! 언제 수업 가능하신가요?',
       };
 
-      await this.socketService.sendMessageToBothUser(
+      await this.chattingService.sendMessageToBothUser(
         studentId,
         userId,
         chatRoomId,
         'problem-image',
         JSON.stringify(problemMessage),
       );
-      await this.socketService.sendMessageToBothUser(
+      await this.chattingService.sendMessageToBothUser(
         studentId,
         userId,
         chatRoomId,
@@ -141,7 +143,7 @@ export class OfferService {
         startTime: startTime.toISOString(),
       };
 
-      await this.socketService.sendMessageToBothUser(
+      await this.chattingService.sendMessageToBothUser(
         userId,
         chatting.teacherId,
         chattingId,
@@ -164,7 +166,7 @@ export class OfferService {
             teacherChatId,
             ChattingStatus.declined,
           );
-          await this.socketService.sendMessageToBothUser(
+          await this.chattingService.sendMessageToBothUser(
             userId,
             offerTeacherId,
             teacherChatId,
